jpa repo + service

+++ src/main/java/com/movieservice/movieserivice/repository/MovieRepo.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface MovieRepo extends JpaRepository<Movie, Integer> {
+public interface MovieRepo extends JpaRepository<Movie, Long> {
 
 }
